+++ src/Models/Traits/AssetParentTrait.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hirtz\Media\Models\Traits;
 
-use Hirtz\Media\helpers\Sizes;
+use Hirtz\Media\Helpers\Sizes;
 use Hirtz\Skeleton\Models\Traits\TypeAttributeTrait;
 use yii\helpers\Inflector;