+++ src/Modules/Admin/Widgets/Buttons/FileUploadButton.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hirtz\Media\Modules\Admin\Widgets\Buttons;
 
-use Hirtz\Media\helpers\Html;
+use Hirtz\Media\Helpers\Html;
 use Hirtz\Media\Models\File;
 use Hirtz\Media\Modules\ModuleTrait;
 use Hirtz\Skeleton\Html\Icon;