+++ src/Modules/Admin/Widgets/Forms/Fields/FilePreviewField.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hirtz\Media\Modules\Admin\Widgets\Forms\Fields;
 
-use Hirtz\Media\helpers\AspectRatio;
+use Hirtz\Media\Helpers\AspectRatio;
 use Hirtz\Media\Models\File;
 use Hirtz\Skeleton\Html\Div;
 use Hirtz\Skeleton\Html\Img;